% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{cbindToColumn}
\alias{cbindToColumn}
\title{Utility function.}
\usage{
cbindToColumn(x, data = NULL, values = NULL)
}
\arguments{
\item{x}{A cbind-vector to be split in columns.}

\item{data}{(Optional) the data frame.}

\item{values}{A vector with new column names.}
}
\value{
A data.frame with
}
\description{
Function to change cbind columns into separated columns.
}
\section{Note}{

Will be moved to other package.
# simulate some gaze data:
dat <- data.frame(
 Subject = rep(1:3, 500),
 Timestamp = rep(1:500, 3),
 AOI = rep( rep( c('other','competitor', 'target'), 3),
 c(184, 172, 144, 51, 264, 185, 127, 2, 371)) )
# add missing data:
dat[sample(nrow(dat), size=15),]$AOI <- NA

# add timebins:
dat$Timebin <- timeBins(dat$Timestamp, 100)

# calculate counts:
c1 <- getCountData('AOI', data=dat, split_by=c('Subject', 'Timebin'))
head(c1)

# make columns:
c1 <- cbindToColumn("AOI", data=c1)
head(c1)
}
\author{
Jacolien van Rij
}
\seealso{
\code{\link[base]{table}}

Other Functions for binomial count data.: \code{\link{countValues}}
}

