% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{convertToBinom}
\alias{convertToBinom}
\title{Utility function.}
\usage{
convertToBinom(success, failure, data = NULL, name = "cbind")
}
\arguments{
\item{success}{Vector with column(s) of count data
defining the number of successes.}

\item{failure}{Vector with column(s) of count data
defining the number of failures.}

\item{data}{Data frame.}

\item{name}{Text string, name of column with the cbind of
successes and failures. Only if \code{data} is specified.}
}
\value{
cbind count data, which sums up to 2 for each row.
}
\description{
Create cbind of successes and failures.
}
\author{
Jacolien van Rij
}
\seealso{
Other Functions for binomial count data.: \code{\link{countValues}};
  \code{\link{getCountData}}; \code{\link{reduceCounts}}
}

