% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{countValues}
\alias{countValues}
\title{Utility function.}
\usage{
countValues(x, values, incl_na = FALSE)
}
\arguments{
\item{x}{A vector to be counted.}

\item{values}{A vector with all possible group names.}

\item{incl_na}{Logical: whether or not to return a count of missing values.}
}
\value{
A table with count information.
}
\description{
Function uses \code{\link[base]{table}} with factors
to count the occurrances of each value in the predictor.
}
\section{Note}{

Values that are not specified in \code{values} will be ignored.
}
\seealso{
\code{\link[base]{table}}

Other Functions for binomial count data.: \code{\link{convertToBinom}};
  \code{\link{getCountData}}; \code{\link{reduceCounts}}
}

