% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeg}
\alias{eeg}
\title{Raw EEG data, single trial, 50Hz.}
\format{
A data frame with 1504 rows and 5 variables:
\describe{
  \item{\code{Electrode}}{Electrode that recorded the EEG.}
  \item{\code{Time}}{Time, time measure from onset of the stimulus.}
  \item{\code{Ampl}}{EEG amplitude, recorded by 32 electrodes.}
  \item{\code{X}}{Approximation of electrode position, relative to Cz. 
Left is negative.}
  \item{\code{Y}}{Approximation of electrode position, relative to Cz. 
Back is negative.}
}
}
\usage{
eeg
}
\description{
A dataset containing a single EEG trial.
}
\author{
Jacolien van Rij
}
\keyword{datasets}
