% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{findAbsMin}
\alias{findAbsMin}
\title{Return the value (or the element with the value) closest to zero.}
\usage{
findAbsMin(x, element = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{element}{Logical: whether or not to return the value (FALSE, default) 
or the index (TRUE).}
}
\value{
The value or index of the element closest to zero (absolute 
minimum).
}
\description{
Return the value (or the element with the value) closest to zero.
}
\examples{
(test <- seq(-25,25, by=3))
min(test[test>0])
max(test[test<0])
min(abs(test))
findAbsMin(test)
}
\author{
Jacolien van Rij
}
\seealso{
Other Utility functions: \code{\link{convertNonAlphanumeric}},
  \code{\link{diff_terms}}, \code{\link{find_difference}},
  \code{\link{find_n_neighbors}}, \code{\link{getDec}},
  \code{\link{getRange}}, \code{\link{group_sort}},
  \code{\link{list2str}}, \code{\link{missing_est}},
  \code{\link{move_n_point}}, \code{\link{print_summary}},
  \code{\link{se}}, \code{\link{summary_data}},
  \code{\link{timeBins}}
}

