% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getDec}
\alias{getDec}
\title{Return the number of decimal places.}
\usage{
getDec(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Number of decimals
}
\description{
Return the number of decimal places.
}
\author{
Based on http://stackoverflow.com/questions/5173692/how-to-return-number-of-decimal-places-in-r
}
\seealso{
Other Utility functions: \code{\link{convertNonAlphanumeric}},
  \code{\link{diff_terms}}, \code{\link{findAbsMin}},
  \code{\link{find_difference}},
  \code{\link{find_n_neighbors}}, \code{\link{getRange}},
  \code{\link{group_sort}}, \code{\link{list2str}},
  \code{\link{missing_est}}, \code{\link{move_n_point}},
  \code{\link{print_summary}}, \code{\link{se}},
  \code{\link{summary_data}}, \code{\link{timeBins}}
}

