% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_utility_functions.R
\name{horiz_error}
\alias{horiz_error}
\title{Utility function.}
\usage{
horiz_error(y, mean, se, minmax = NULL, ...)
}
\arguments{
\item{y}{Vector with y-values.}

\item{mean}{Vector with means.}

\item{se}{Vector with errors or confidence bands.}

\item{minmax}{Optional argument, vector with two values indicating the
minimum and maximum value for the error bars. If NULL (default) the error
bars are not corrected.}

\item{...}{Optional graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
Add horizontal error bars.
}
\author{
Jacolien van Rij
}
\seealso{
Other Utility functions for plotting: \code{\link{addInterval}};
  \code{\link{alphaPalette}}; \code{\link{alpha}};
  \code{\link{dotplot_error}}; \code{\link{emptyPlot}};
  \code{\link{errorBars}}; \code{\link{fadeRug}};
  \code{\link{fill_area}}; \code{\link{find_difference}};
  \code{\link{getCoords}}; \code{\link{gradientLegend}};
  \code{\link{plot_error}}
}

