% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itsadug.R
\name{info}
\alias{info}
\title{Information on how to cite this package}
\usage{
info(input = NULL)
}
\arguments{
\item{input}{Optional parameter. Normally (NULL) the citation info is 
printed. If value 'version' then only the version is printed.}
}
\description{
Information on how to cite this package
}
\examples{
info()
info('version')
citation(package='itsadug')
# To get info about R version:
R.version.string
}
\seealso{
\code{\link[utils]{citation}}, \code{\link[base]{R.version}},
\code{\link[utils]{sessionInfo}}

Other Functions for package use: 
\code{\link{infoMessages}()}
}
\concept{Functions for package use}
