% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list2str}
\alias{list2str}
\title{Combine list values as string.}
\usage{
list2str(x, inputlist)
}
\arguments{
\item{x}{A vector with the names or numbers of list elements to be combined.}

\item{inputlist}{A (named) list with information, e.g., graphical parameter settings.}
}
\value{
String
}
\description{
Combine list values as string.
}
\seealso{
Other Utility functions: \code{\link{convertNonAlphanumeric}},
  \code{\link{diff_terms}}, \code{\link{findAbsMin}},
  \code{\link{find_difference}},
  \code{\link{find_n_neighbors}}, \code{\link{getDec}},
  \code{\link{getRange}}, \code{\link{group_sort}},
  \code{\link{missing_est}}, \code{\link{move_n_point}},
  \code{\link{print_summary}}, \code{\link{se}},
  \code{\link{summary_data}}, \code{\link{timeBins}}
}

