% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_summary}
\alias{print_summary}
\title{Print a named list of strings, output from \code{\link{summary_data}}.}
\usage{
print_summary(sumlist, title = NULL)
}
\arguments{
\item{sumlist}{Named list, output of \code{\link{summary_data}}.}

\item{title}{Optional, text string that will be printed as title.}
}
\description{
Print a named list of strings, output from \code{\link{summary_data}}.
}
\seealso{
Other Utility functions: 
\code{\link{convertNonAlphanumeric}()},
\code{\link{corfit}()},
\code{\link{diff_terms}()},
\code{\link{find_difference}()},
\code{\link{missing_est}()},
\code{\link{modeledf}()},
\code{\link{observations}()},
\code{\link{refLevels}()},
\code{\link{res_df}()},
\code{\link{summary_data}()},
\code{\link{timeBins}()}
}
\author{
Jacolien van Rij
}
\concept{Utility functions}
