% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{reduceCounts}
\alias{reduceCounts}
\title{Reducing cbind binomial count data to avoid autocorrelation issues.}
\usage{
reduceCounts(x, y = NULL)
}
\arguments{
\item{x}{cbind data column (leave \code{y} NULL), or a vector
specify \code{y}).}

\item{y}{Vector, in case \code{x} is vector, otherwise NULL.}
}
\value{
cbind count data, which sums up to 2 for each row.
}
\description{
Reducing cbind binomial count data to avoid autocorrelation issues.
}
\author{
Jacolien van Rij
}
\seealso{
Other Functions for binomial count data.: \code{\link{convertToBinom}};
  \code{\link{countValues}}; \code{\link{getCountData}}
}

