% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{refLevels}
\alias{refLevels}
\title{Return a list with reference levels for each factor.}
\usage{
refLevels(data)
}
\arguments{
\item{data}{Data}
}
\value{
Named list with reference levels for each factor.
}
\description{
Function for retrieving all reference levels.
}
\seealso{
Other Utility functions: 
\code{\link{convertNonAlphanumeric}()},
\code{\link{corfit}()},
\code{\link{diff_terms}()},
\code{\link{find_difference}()},
\code{\link{missing_est}()},
\code{\link{modeledf}()},
\code{\link{observations}()},
\code{\link{print_summary}()},
\code{\link{res_df}()},
\code{\link{summary_data}()},
\code{\link{timeBins}()}
}
\author{
Jacolien van Rij
}
\concept{Utility functions}
