% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{se}
\alias{se}
\title{Calculate standard error of the mean.}
\usage{
se(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Standard Error of the mean.
}
\description{
Calculate standard error of the mean.
}
\seealso{
Other Utility functions: \code{\link{convertNonAlphanumeric}},
  \code{\link{diff_terms}}, \code{\link{findAbsMin}},
  \code{\link{find_difference}},
  \code{\link{find_n_neighbors}}, \code{\link{getDec}},
  \code{\link{getRange}}, \code{\link{group_sort}},
  \code{\link{list2str}}, \code{\link{missing_est}},
  \code{\link{move_n_point}}, \code{\link{print_summary}},
  \code{\link{summary_data}}, \code{\link{timeBins}}
}

