% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Simulated time series data.}
\format{
A data frame with 75600 rows and 6 variables:
\describe{
  \item{\code{Group}}{Age group of participants: Adults or Children.}
  \item{\code{Time}}{Time, time measure from start of each time series.}
  \item{\code{Trial}}{Trial in the experiment, centered around zero.}
  \item{\code{Condition}}{Continuous variable, ranging from -1 to 4. 
For example, stimulus onset asynchrony.}
  \item{\code{Subject}}{Code for individual participants.}
  \item{\code{Y}}{Time series measure. 
Similar to pupil size, sensor position, or voltage.}
}
}
\usage{
simdat
}
\description{
A dataset containing the sine wave data with random noise added.
}
\author{
Jacolien van Rij
}
\keyword{datasets}
