% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary_data}
\alias{summary_data}
\title{Print a descriptive summary of a data frame.}
\usage{
summary_data(data, print = TRUE, n = 10)
}
\arguments{
\item{data}{A data frame.}

\item{print}{Logical: whether or not to print the summary.}

\item{n}{Number: maximum number of values being mentioned in the summary.
If NULL all values are being mentioned. Defaults to 10.}
}
\value{
Optionally returns a named list with info.
}
\description{
The function prints a summary of the data. 
Similar to the function \code{\link[utils]{str}}, but easier readable.
}
\examples{
data(simdat)
summary_data(simdat)
}
\seealso{
Other Utility functions: 
\code{\link{convertNonAlphanumeric}()},
\code{\link{corfit}()},
\code{\link{diff_terms}()},
\code{\link{find_difference}()},
\code{\link{missing_est}()},
\code{\link{modeledf}()},
\code{\link{observations}()},
\code{\link{print_summary}()},
\code{\link{refLevels}()},
\code{\link{res_df}()},
\code{\link{timeBins}()}
}
\author{
Jacolien van Rij
}
\concept{Utility functions}
