% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_utility_functions.R
\name{alpha}
\alias{alpha}
\title{Utility function.}
\usage{
alpha(x, f = 0.5)
}
\arguments{
\item{x}{A color or a vector with color values.}

\item{f}{A number for adjusting the transparency ranging from 0 (completely
transparent) to 1 (not transparent).}
}
\description{
Wrapper around \code{\link[grDevices]{adjustcolor}}.
}
\section{Note}{

Does not always work for x11 panels.
}
\examples{
emptyPlot(100,100, h=50, v=50)
rect(25,25,75,75, col=alpha('red',f=1))
rect(35,41,63,81, col=alpha(rgb(0,1,.5),f=.25),
   border=alpha(rgb(0,1,.5), f=.65), lwd=4)
}
\seealso{
Other Utility functions for plotting: \code{\link{addInterval}};
  \code{\link{alphaPalette}}; \code{\link{color_contour}};
  \code{\link{dotplot_error}}; \code{\link{emptyPlot}};
  \code{\link{errorBars}}; \code{\link{fadeRug}};
  \code{\link{fill_area}}; \code{\link{find_difference}};
  \code{\link{getCoords}}; \code{\link{getFigCoords}};
  \code{\link{gradientLegend}}; \code{\link{horiz_error}};
  \code{\link{marginDensityPlot}}; \code{\link{plot_error}}
}

