% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_utility_functions.R
\name{alpha}
\alias{alpha}
\title{Utility function.}
\usage{
alpha(x, f = 0.5)
}
\arguments{
\item{x}{A color or a vector with color values.}

\item{f}{A number for adjusting the transparency ranging from 0 (completely
transparent) to 1 (not transparent).}
}
\description{
Wrapper around \code{\link[grDevices]{adjustcolor}}.
}
\seealso{
Other Utility functions for plotting: \code{\link{alphaPalette}};
  \code{\link{dotplot_error}}; \code{\link{emptyPlot}};
  \code{\link{errorBars}}; \code{\link{fadeRug}};
  \code{\link{fill_area}}; \code{\link{find_difference}};
  \code{\link{getCoords}}; \code{\link{gradientLegend}};
  \code{\link{horiz_error}}; \code{\link{plot_error}}
}

