% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inspect_residuals.R
\name{check_resid}
\alias{check_resid}
\title{Inspect residuals of regression models.}
\usage{
check_resid(model, AR_start = NULL, split_by = NULL)
}
\arguments{
\item{model}{A GAMM model, resulting from the functions
\code{\link[mgcv]{gam}} or \code{\link[mgcv]{bam}}.}

\item{AR_start}{ADD DESCRIPTION}

\item{split_by}{ADD DESCRIPTION}
}
\description{
Inspect residuals of regression models.
}
\author{
Jacolien van Rij
}

