% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{diff_terms}
\alias{diff_terms}
\title{Utility unction.}
\usage{
diff_terms(model1, model2)
}
\arguments{
\item{model1}{A fitted regression model (using lm, glm, gam, or bam).}

\item{model2}{A fitted regression model (using lm, glm, gam, or bam).}
}
\value{
A list with model terms that are not shared by both models.
}
\description{
Utility unction.
}
\examples{
data(simdat)

# Fit simple GAM model:
gam1 <- bam(Y ~ s(Time), data=simdat)
gam2 <- bam(Y ~ Group+s(Time), data=simdat)
diff_terms(gam1, gam2)
}
\author{
Jacolien van Rij
}
\seealso{
Other utility functions: \code{\link{missing_est}}
}

