% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_utility_functions.R
\name{errorBars}
\alias{errorBars}
\title{Utility function.}
\usage{
errorBars(x, mean, se, minmax = NULL, ...)
}
\arguments{
\item{x}{Vector with x-values.}

\item{mean}{Vector with means.}

\item{se}{Vector with errors or confidence bands.}

\item{minmax}{Optional argument, vector with two values indicating the
minimum and maximum value for the error bars. If NULL (default) the error
bars are not corrected.}

\item{...}{Optional graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
Add vertical error bars.
}
\author{
Jacolien van Rij
}
\seealso{
Other Utility functions for plotting: \code{\link{addInterval}};
  \code{\link{alphaPalette}}; \code{\link{alpha}};
  \code{\link{color_contour}}; \code{\link{dotplot_error}};
  \code{\link{emptyPlot}}; \code{\link{fadeRug}};
  \code{\link{fill_area}}; \code{\link{find_difference}};
  \code{\link{getCoords}}; \code{\link{getFigCoords}};
  \code{\link{gradientLegend}}; \code{\link{horiz_error}};
  \code{\link{marginDensityPlot}}; \code{\link{plot_error}}
}

