% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_difference.R
\name{find_difference}
\alias{find_difference}
\title{Find the regions in which the smooth is significantly different from zero.}
\usage{
find_difference(mean, se, f = 1.96, xVals = NULL)
}
\arguments{
\item{mean}{A vector with smooth predictions.}

\item{se}{A vector with the standard error on the smooth predictions.}

\item{f}{A number to multiply the \code{se} with, to convert the \code{se}
into confidence intervals. Defaults to 1.96 (95\% CI), use 2.58 for 99\%CI.}

\item{xVals}{Optional vector with x values for the smooth.
When \code{xVals} is provided, the regions are returned in terms of x-
values, otherwise as indices.}
}
\value{
The function returns a list with start points of each region
(\code{start}) and end points of each region (\code{end}). The logical
\code{xVals} indicates whether the returned values are on the x-scale
(TRUE) or indices (FALSE).
}
\description{
Find the regions in which the smooth is significantly different from zero.
}
\author{
Jacolien van Rij
}
\seealso{
Other Utility functions for plotting: \code{\link{alphaPalette}};
  \code{\link{alpha}}; \code{\link{dotplot_error}};
  \code{\link{emptyPlot}}; \code{\link{errorBars}};
  \code{\link{fadeRug}}; \code{\link{fill_area}};
  \code{\link{getCoords}}; \code{\link{gradientLegend}};
  \code{\link{horiz_error}}; \code{\link{plot_error}}
}

