% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{find_n_neighbors}
\alias{find_n_neighbors}
\title{Utility function.}
\usage{
find_n_neighbors(el, n, max)
}
\arguments{
\item{el}{A numeric vector.}

\item{n}{Number indicating how many points around the elements of \code{el}
need to be selected.}

\item{max}{The maximum value of the returned elements.}
}
\value{
A vector with the elements of x surrounded by n points.
}
\description{
Utility function.
}
\author{
Jacolien van Rij
}
\seealso{
Other Data utility functions: \code{\link{findAbsMin}};
  \code{\link{group_sort}}; \code{\link{move_n_point}};
  \code{\link{summary_data}}
}

