% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{findAbsMin}
\alias{findAbsMin}
\title{Utility function.}
\usage{
findAbsMin(x, element = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{element}{Logical: whether or not to return the value (FALSE, default)
or the index (TRUE).}
}
\value{
The value or index of the element closest to zero (absolute
minimum).
}
\description{
Utility function.
}
\examples{
(test <- seq(-25,25, by=3))
min(test[test>0])
max(test[test<0])
min(abs(test))
findAbsMin(test)
}
\author{
Jacolien van Rij
}
\seealso{
Other Data utility functions: \code{\link{find_n_neighbors}};
  \code{\link{group_sort}}; \code{\link{move_n_point}};
  \code{\link{print_summary}}; \code{\link{summary_data}}
}

