% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{gamtabs}
\alias{gamtabs}
\title{Convert the summary  into a Latex table.}
\usage{
gamtabs(model, caption = " ", label = "tab.gam", pnames = NA,
  snames = NA, ptab = NA, stab = NA, ...)
}
\arguments{
\item{model}{A GAM(M) model build in the package \code{\link[mgcv]{mgcv}}
using \code{\link[mgcv]{gam}} or \code{\link[mgcv]{bam}}.}

\item{caption}{A string with the caption for the table.}

\item{label}{A string for the label to refer to the table in the markdown
document.}

\item{pnames}{A vector with labels to relabel the rows in the parametric
part of the summary.}

\item{snames}{A vector with labels to relabel the rows in the smooth
part of the summary.}

\item{ptab}{A vector with labels to relabel the column names of the
parametric summary.}

\item{stab}{A vector with labels to relabel the column names of the
smooth summary.}

\item{...}{Optional additional arguments which are passed to
\code{xtable} (see 'help(xtable)').}
}
\value{
A vector with color values.
}
\description{
Convert the summary  into a Latex table.
}
\section{Note}{

This function is useful for markdown documents using the package
\code{knitr} to integrate R code with Latex and Sweave. This
function requires the package \code{xtable}.
}
\author{
R. Harald Baayen
}
\seealso{
\code{\link[mgcv]{summary.gam}}, \code{\link[mgcv]{gam}},
\code{\link[mgcv]{bam}}.
}

