% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_utility_functions.R
\name{getCoords}
\alias{getCoords}
\title{Utility function.}
\usage{
getCoords(pos = 1.1, side = 1)
}
\arguments{
\item{pos}{A number indicating the proportion on the x-axis. Default is 1.1.}

\item{side}{Which axis to choose: 1=bottom, 2=left, 3=top, 4=right. Default is 1.}
}
\description{
Function for positioning a legend or label in or outside the
plot region based on proportion of the plot region rather than Cartesian
coordinates.
}
\author{
Jacolien van Rij
}
\seealso{
Other Utility functions for plotting: \code{\link{addInterval}};
  \code{\link{alphaPalette}}; \code{\link{alpha}};
  \code{\link{dotplot_error}}; \code{\link{emptyPlot}};
  \code{\link{errorBars}}; \code{\link{fadeRug}};
  \code{\link{fill_area}}; \code{\link{find_difference}};
  \code{\link{gradientLegend}}; \code{\link{horiz_error}};
  \code{\link{plot_error}}
}

