% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{getCountData}
\alias{getCountData}
\title{Reducing fixations to count data.}
\usage{
getCountData(x, split_by, data = NULL, values = NULL, sumvalues = NULL,
  incl_na = FALSE, cbind = FALSE)
}
\arguments{
\item{x}{Name of a column in \code{data} or vector with values
to be counted.}

\item{split_by}{Vector with column names in \code{data} or named list with
grouping predictors.}

\item{data}{Optional: data frame.}

\item{values}{Values of \code{x} that should be counted.
If NULL (default) all values are counted.}

\item{sumvalues}{Named list that specifies wchich counts can be combined
 (summed). The values listed in \code{sumvalues} must be specified in
 \code{values}. If \code{incl_na} is TRUE, \code{sumvalues} can also
include 'NA' (with quotes).}

\item{incl_na}{Logical: whether or not to return a count of missing values.}

\item{cbind}{Logical: whether or not to include the counts as cbinded
values. Defaults to FALSE, i.e., each countwill be included as new column.}
}
\value{
A data frame with cbinded count information.
}
\description{
Function uses \code{\link[base]{table}} with factors
to count the occurrances of each value in the predictor.
}
\section{Note}{

Values that are not specified in \code{values} will be ignored.
}
\examples{
# simulate some gaze data:
dat <- data.frame(
 Subject = rep(1:3, 500),
 Timestamp = rep(1:500, 3),
 AOI = rep( rep( c('other','competitor', 'target'), 3),
 c(184, 172, 144, 51, 264, 185, 127, 2, 371)) )
# add missing data:
dat[sample(nrow(dat), size=15),]$AOI <- NA

# add timebins:
dat$Timebin <- timeBins(dat$Timestamp, 100)

# calculate counts:
c1 <- getCountData('AOI', data=dat, split_by=c('Subject', 'Timebin'))
head(c1)
colnames(c1)
# calculating proportions:
c1$prop <- c1$AOI.target / ( c1$AOI.competitor+c1$AOI.other )

# Using cbinds:
c2 <- getCountData('AOI', data=dat, split_by=c('Subject', 'Timebin'),
     cbind=FALSE)
head(c2)
colnames(c2)
# calculating proportions:
c2$prop <- c2$AOI[,'target'] / ( c2$AOI[,'competitor']+c2$AOI[,'other'])

# calculating counts for specific values, including missing data.
# Note that 'distractor' is not in the data:
c3 <- getCountData('AOI', data=dat, split_by=c('Subject', 'Timebin'),
values=c('target', 'distractor', 'competitor', 'other'), incl_na=TRUE)
head(c3)

# summing up specific values:
c4 <- getCountData('AOI', data=dat, split_by=c('Subject', 'Timebin'),
values=levels(dat$AOI),
sumvalues=list('correct'=c('target', 'competitor'),
'incorrect'=c('other','NA')), incl_na=TRUE, cbind=TRUE)
head(c4)
names(c4)
# or as separate columns:
c4 <- getCountData('AOI', data=dat, split_by=c('Subject', 'Timebin'),
values=levels(dat$AOI),
sumvalues=list('correct'=c('target', 'competitor'),
'incorrect'=c('other','NA')), incl_na=TRUE)
head(c4)
names(c4)
}
\author{
Jacolien van Rij
}
\seealso{
\code{\link[base]{table}}

Other Functions for gaze data: \code{\link{timeBins}}
}

