% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{getCountData}
\alias{getCountData}
\title{Utility function.}
\usage{
getCountData(x, split_by, data = NULL, values = NULL, incl_na = FALSE)
}
\arguments{
\item{x}{Name of a column in \code{data} or vector with values
to be counted.}

\item{split_by}{Vector with column names in \code{data} or named list with
grouping predictors.}

\item{data}{Optional: data frame.}

\item{values}{Values of \code{x} that should be counted.
If NULL (default) all values are counted.}

\item{incl_na}{Logical: whether or not to return a count of missing values.}
}
\value{
A data frame with cbinded count information.
}
\description{
Function uses \code{\link[base]{table}} with factors
to count the occurrances of each value in the predictor.
}
\section{Note}{

Values that are not specified in \code{values} will be ignored.
}
\seealso{
\code{\link[base]{table}}

Other Functions for binomial count data.: \code{\link{convertToBinom}};
  \code{\link{countValues}}; \code{\link{reduceCounts}}
}

