% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{getDec}
\alias{getDec}
\title{Utility function.}
\usage{
getDec(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Number of decimals
}
\description{
Utility function.
}
\author{
Jacolien van Rij
}
\seealso{
Other Data utility functions: \code{\link{findAbsMin}};
  \code{\link{find_n_neighbors}}; \code{\link{getRange}};
  \code{\link{group_sort}}; \code{\link{list2str}};
  \code{\link{move_n_point}}; \code{\link{print_summary}};
  \code{\link{summary_data}}; \code{\link{timeBins}}
}

