% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{list2str}
\alias{list2str}
\title{Utility function.}
\usage{
list2str(x, inputlist)
}
\arguments{
\item{x}{A vector with the names or numbers of list elements to be combined.}

\item{inputlist}{A (named) list with information, e.g., graphical parameter settings.}
}
\value{
String
}
\description{
Combine list values as string.
}
\seealso{
Other Data utility functions: \code{\link{findAbsMin}};
  \code{\link{find_n_neighbors}}; \code{\link{getDec}};
  \code{\link{getRange}}; \code{\link{group_sort}};
  \code{\link{move_n_point}}; \code{\link{print_summary}};
  \code{\link{summary_data}}; \code{\link{timeBins}}
}

