% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{missing_est}
\alias{missing_est}
\title{Utility unction.}
\usage{
missing_est(model)
}
\arguments{
\item{model}{A fitted regression model (using lm, glm, gam, or bam).}
}
\value{
The indices of the data that were not fitted by the model.
}
\description{
Utility unction.
}
\author{
Jacolien van Rij
}

