% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{missing_est}
\alias{missing_est}
\title{Utility unction.}
\usage{
missing_est(model)
}
\arguments{
\item{model}{A fitted regression model (using lm, glm, gam, or bam).}
}
\value{
The indices of the data that were not fitted by the model.
}
\description{
Utility unction.
}
\examples{
data(simdat)

# Add missing values:
set.seed(123)
simdat[sample(nrow(simdat), size=20),]$Y <- NA
# Fit simple linear model:
lm1 <- lm(Y ~ Time, data=simdat)
na.el <- missing_est(lm1)
length(na.el)
}
\author{
Jacolien van Rij
}
\seealso{
Other utility functions: \code{\link{convertNonAlphanumeric}};
  \code{\link{diff_terms}}
}

