% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plotDiff}
\alias{plot_diff}
\title{Plot difference curve based on model predictions.}
\usage{
plot_diff(model, xvar, by_predictor, comp_levels = NULL, eegAxis = F,
  ylim = NULL, ylab = NULL, main = NULL, plot = TRUE, ...)
}
\arguments{
\item{model}{A GAMM model, resulting from the functions
\code{\link[mgcv]{gam}} or \code{\link[mgcv]{bam}}.}

\item{xvar}{Name of continuous predictor that should be plotted on the x-
axis.}

\item{by_predictor}{Name of the grouping predictor (categorical variable).}

\item{comp_levels}{Two levels of \code{by_predictor} to calculate the
difference for.}

\item{eegAxis}{Logical: whether or not to reverse the y-axis, plotting
negative values upwards. Default is FALSE.}

\item{ylim}{Range of y-axis. If not specified, the function automatically
generates an appropriate y-axis.}

\item{ylab}{Text string, alternative label for y-axis.}

\item{main}{Text string, alternative title for plot.}

\item{plot}{Logical: whether or not to plot the difference. If FALSE, then
the output is returned as a list, with the estimated difference
(\code{est}) and the standard error over the estimate (\code{se.est}) and
the x-values (\code{x}). Default is TRUE.}

\item{...}{Optional arguments for plot.}
}
\value{
If the result is not being plotted, a list is
returned with the estimated difference (\code{est}) and the standard error
over the estimate (\code{se.est}) and the x-values (\code{x}) is returned.
}
\description{
Plot difference curve based on model predictions.
}
\author{
Martijn Wieling
}
\seealso{
Other functions for interpreting nonlinear effects: \code{\link{fvisgam}},
  \code{\link{vis.gam2}}; \code{\link{plotDiff2D}},
  \code{\link{plot_diff2}}; \code{\link{plot_parametric}};
  \code{\link{plot_smooth}}; \code{\link{pvis.gam}},
  \code{\link{pvisgam}}
}

