% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff2}
\alias{plotDiff2D}
\alias{plot_diff2}
\title{Plot difference surface based on model predictions.}
\usage{
plot_diff2(model, xvar, yvar, by_predictor, comp_levels = NULL, plotCI = F,
  color = "topo", nCol = 100, col = NULL, add.color.legend = TRUE,
  n.grid = 30, nlevels = 10, zlim = NULL, main = NULL)
}
\arguments{
\item{model}{A GAMM model, resulting from the functions
\code{\link[mgcv]{gam}} or \code{\link[mgcv]{bam}}.}

\item{xvar}{Name of continuous predictor that should be plotted on the x-
axis.}

\item{yvar}{Name of continuous predictor that should be plotted on the y-
axis.}

\item{by_predictor}{Name of the grouping predictor (categorical variable).}

\item{comp_levels}{Two levels of \code{by_predictor} to calculate the
difference for.}

\item{plotCI}{Logical: whether or not to plot confidence intervals.}

\item{color}{Colorpalette}

\item{nCol}{Range of colors of background of contour plot.}

\item{col}{Line color.}

\item{add.color.legend}{Logical: whether or not to add a color legend.
Default is TRUE. If FALSE (omitted), one could use the function
\code{\link{gradientLegend}} to add a legend manually at any position.}

\item{n.grid}{Resolution.}

\item{nlevels}{Levels of contour lines.}

\item{zlim}{A two item array giving the lower and upper limits for the z-
axis scale. NULL to choose automatically.}

\item{main}{Title of plot.}
}
\value{
If the result is not being plotted, a list is
returned with the estimated difference (\code{est}) and the standard error
over the estimate (\code{se.est}) and the x-values (\code{x}) is returned.
}
\description{
Plot difference surface based on model predictions.
}
\examples{
# see the vignette for examples:
vignette("plotfunctions", package="itsadug")
}
\author{
Martijn Wieling
}
\seealso{
Other functions for interpreting nonlinear effects: \code{\link{fvisgam}},
  \code{\link{vis.gam2}}; \code{\link{plotDiff}},
  \code{\link{plot_diff}}; \code{\link{plot_parametric}};
  \code{\link{plot_smooth}}; \code{\link{pvis.gam}},
  \code{\link{pvisgam}}
}

