% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basic_plots.R
\name{plot_error}
\alias{plot_error}
\title{Utility function}
\usage{
plot_error(x, fit, se.fit, shade = FALSE, f = 1, col = "black",
  border = NA, alpha = 0.25, ...)
}
\arguments{
\item{x}{Vector with values on x-axis.}

\item{fit}{Vector with values on y-axis.}

\item{se.fit}{Vector with standard error.}

\item{shade}{Logical: whether or not to produce shaded regions as
confidence bands.}

\item{f}{Factor for converting standard error in confidence intervals.
Defaults to 1. Use 1.96 for 95\% CI, and 2.58 for 99\% CI.}

\item{col}{Color for lines and confindence bands.}

\item{border}{A number, indicating the linewidth of the border around
shaded area. No border with value NA (default). Only applies when
\code{shade=TRUE}.}

\item{alpha}{Transparency of shaded area. Number between 0
(completely transparent) and 1 (not transparent).}

\item{...}{Optional arguments for the lines. See \code{\link{par}}.}
}
\description{
Plot line with confidence intervals.
}
\examples{
# see the vignette on visualization,
# section 'Using predictions to customize plots', for an example:
vignette(topic="plotfunctions")
}
\author{
Jacolien van Rij
}
\seealso{
Other Utility functions for plotting: \code{\link{alphaPalette}};
  \code{\link{alpha}}; \code{\link{dotplot_error}};
  \code{\link{emptyPlot}}; \code{\link{errorBars}};
  \code{\link{fadeRug}}; \code{\link{fill_area}};
  \code{\link{find_difference}}; \code{\link{getCoords}};
  \code{\link{gradientLegend}}; \code{\link{horiz_error}}
}

