% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_summary}
\alias{print_summary}
\title{Print a named list of strings, output from \code{\link{summary_data}}.}
\usage{
print_summary(sumlist, title = NULL)
}
\arguments{
\item{sumlist}{Named list, output of \code{\link{summary_data}}.}

\item{title}{Optional, text string that will be printed as title.}
}
\description{
Print a named list of strings, output from \code{\link{summary_data}}.
}
\author{
Jacolien van Rij
}
\seealso{
Other Utility functions: \code{\link{convertNonAlphanumeric}},
  \code{\link{diff_terms}}, \code{\link{findAbsMin}},
  \code{\link{find_difference}},
  \code{\link{find_n_neighbors}}, \code{\link{getDec}},
  \code{\link{getRange}}, \code{\link{group_sort}},
  \code{\link{list2str}}, \code{\link{missing_est}},
  \code{\link{move_n_point}}, \code{\link{se}},
  \code{\link{summary_data}}, \code{\link{timeBins}}
}

