% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{print_summary}
\alias{print_summary}
\title{Utility function.}
\usage{
print_summary(sumlist, title = NULL)
}
\arguments{
\item{sumlist}{Named list, output of \code{\link{summary_data}}.}

\item{title}{Optional, text string that will be printed as title.}
}
\description{
Utility function.
}
\author{
Jacolien van Rij
}
\seealso{
Other Data utility functions: \code{\link{findAbsMin}};
  \code{\link{find_n_neighbors}}; \code{\link{group_sort}};
  \code{\link{move_n_point}}; \code{\link{summary_data}};
  \code{\link{timeBins}}
}

