% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resid_gam.R
\name{resid_gam}
\alias{resid.gam}
\alias{resid_gam}
\title{Extract model residuals and remove the autocorrelation accounted for.}
\usage{
resid_gam(model, AR_start = NULL, incl_na = F, return_all = F)
}
\arguments{
\item{model}{A GAMM model build with \code{\link[mgcv]{gam}} or
\code{\link[mgcv]{bam}}.}

\item{AR_start}{Optional: vector with logicals, indicating the start of
events.
Default is NULL, because generally the function can retrieve all necessary
information from the model.}

\item{incl_na}{Whether or not to include missing values (NA)when returning
the residuals. Defaults to FALSE.}

\item{return_all}{Default is FALSE. Returns a list with normal residuals,
corrected residuals, and the value of rho.}
}
\value{
Corrected residuals.
}
\description{
Extract model residuals and remove the autocorrelation accounted for.
}
\author{
Jacolien van Rij
}
\seealso{
\code{\link[stats]{resid}}

Other functions for model criticism: \code{\link{acf.n.plots}},
  \code{\link{acf_n_plots}}; \code{\link{acf.plot}},
  \code{\link{acf_plot}}; \code{\link{check_normaldist}}
}

