% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{se}
\alias{se}
\title{Calculate standard error of the mean.}
\usage{
se(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Standard Error of the mean.
}
\description{
Calculate standard error of the mean.
}
\author{
Jacolien van Rij
}

