% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_utility_functions.R
\name{summary_data}
\alias{summary_data}
\title{Utility function.}
\usage{
summary_data(data, print = TRUE, n = 10)
}
\arguments{
\item{data}{A data frame.}

\item{print}{Logical: whether or not to print the summary.}

\item{n}{Number: maximum number of values being mentioned in the summary.
If NULL all values are being mentioned. Defaults to 10.}
}
\value{
Optionally returns a named list with info.
}
\description{
The function prints a summary of the data.
Similar to the function \code{\link[utils]{str}}, but easier readable.
}
\examples{
data(simdat)
summary_data(simdat)
}
\author{
Jacolien van Rij
}
\seealso{
Other Data utility functions: \code{\link{findAbsMin}};
  \code{\link{find_n_neighbors}}; \code{\link{group_sort}};
  \code{\link{move_n_point}}; \code{\link{print_summary}}
}

